# contracts/openapi/api-v1.json
{
  "components": {
    "schemas": {
      "ApiErrorDetailDto": {
        "properties": {
          "code": {
            "example": "invalid_request",
            "type": "string"
          },
          "field": {
            "example": "fieldName",
            "type": "string"
          }
        },
        "required": ["code"],
        "type": "object"
      },
      "ApiErrorResponseDto": {
        "properties": {
          "code": {
            "example": "NOT_FOUND",
            "type": "string"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ApiErrorDetailDto"
            },
            "type": "array"
          },
          "message": {
            "example": "Resource not found.",
            "type": "string"
          },
          "requestId": {
            "example": "0193f8d7-7f03-7f25-a4c0-f043f3d78a50",
            "type": "string"
          },
          "status": {
            "example": 404,
            "type": "number"
          }
        },
        "required": ["code", "message", "requestId", "status"],
        "type": "object"
      },
...
      "IdentityResponseDto": {
        "properties": {
          "roles": {
            "items": {
              "enum": ["agent", "support_admin", "tenant_admin", "tenant_owner", "viewer"],
              "type": "string"
            },
            "type": "array"
          },
          "subject": {
            "example": "f25c47ed-6f91-469b-865a-84f42b088a7d",
            "type": "string"
          },
          "tenantContext": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": ["roles", "subject", "tenantContext"],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Providerfreie, synthetische Foundation API.",
    "title": "Voice AI Agent API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/identity/me": {
      "get": {
        "operationId": "IdentityController_getIdentity",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponseDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Return the validated identity without a tenant context",
        "tags": ["identity"]
      }
    },
...
  },
  "servers": [],
...
}